Easier, with simple updates
Now only have to select the output folder, so everything is a bit
faster. Code is cleaned up a bit and organized slightly better - more
improvements needed there, however. A R and R^2 matrix is shown for
motion and ICA comps - perhaps someone will find that useful. Finally,
the TSNR plots use the same axis values, so figures are directly
comparable (ignoring the problem of low frequency fluctuations in TSOC
and E2 data).